tests: move Python 2.7 workaround to '_helpers'
Use it in KMS tests, too.
